+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.2.7] - 2024-07-18
+- New `Package.create(name)` static function.
+- New `pkg.latestVersion()` function to return the latest version of a package.
+- New `pkg.toLatestVersion()` function to change the version of a package to the latest version.
+
 ## [0.2.6] - 2024-07-04
 ### Added
 - New `packageUrl` property to get the package landing page.
@@ -62,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2024-04-14
 First version
 
+[0.2.7]: https://github.com/oscarotero/nudd/compare/v0.2.6...v0.2.7
 [0.2.6]: https://github.com/oscarotero/nudd/compare/v0.2.5...v0.2.6
 [0.2.5]: https://github.com/oscarotero/nudd/compare/v0.2.4...v0.2.5
 [0.2.4]: https://github.com/oscarotero/nudd/compare/v0.2.3...v0.2.4